refined code for all relavant page.tsx
refined code for the following page.tsx:
- app/dashboard
- app/policy/add
